Implemented Strategy Pattern for CPP file
